Merging develop

+++ post_processing/pylbo/utilities/automation.py
@@ -14,6 +14,7 @@ from .defaults import precoded_runs, \
 from .exceptions import DictNotEmpty, \
     InconsistentNumberOfRuns
 from ..data_management.file_handler import select_files
+from ..utilities.logger import pylboLogger
 
 LEGOLAS_PAR = (LEGOLAS_OUT / 'parfiles').resolve()
 
@@ -57,6 +58,8 @@ def generate_parfiles(parfile_dict=None, basename_parfile=None, output_dir=None)
         parfile_dict = get_precoded_run(chosen_pr)
     if output_dir is None:
         output_dir = LEGOLAS_PAR
+    if isinstance(output_dir, str):
+        output_dir = Path(output_dir).resolve()
     _check_directories(output_dir)
     nb_runs = parfile_dict.pop('number_of_runs', 1)
     namelist = {}
@@ -67,6 +70,11 @@ def generate_parfiles(parfile_dict=None, basename_parfile=None, output_dir=None)
                                         'remove_spurious_eigenvalues', 'nb_spurious_eigenvalues'])
     update_namelist('savelist', ['write_matrices', 'write_eigenfunctions', 'show_results',
                                  'basename_datfile', 'basename_logfile', 'output_folder', 'logging_level'])
+    update_namelist('physicslist', ['mhd_gamma', 'flow', 'radiative_cooling', 'ncool', 'cooling_curve',
+                                    'external_gravity', 'thermal_conduction', 'use_fixed_tc_para', 'fixed_tc_para_value',
+                                    'use_fixed_tc_perp', 'fixed_tc_perp_value', 'resistivity', 'use_fixed_resistivity',
+                                    'fixed_eta_value'])
+    update_namelist('unitslist', ['cgs_units', 'unit_density', 'unit_temperature', 'unit_magneticfield', 'unit_length'])
     if parfile_dict.get('parameters') is not None:
         namelist.update({'paramlist': parfile_dict.pop('parameters')})
     # we should have popped everything from the dictionary so it should be empty.
@@ -90,7 +98,7 @@ def generate_parfiles(parfile_dict=None, basename_parfile=None, output_dir=None)
         run_prepended = "{:04d}".format(run)
         if nb_runs == 1:
             run_prepended = ''
-        parfile_dict = {'gridlist': {}, 'equilibriumlist': {}, 'paramlist': {}, 'savelist': {}}
+        parfile_dict = {key: {} for key in namelist.keys()}
         # create dictionary for single run
         for name, nl_dict in namelist.items():
             for key, item in nl_dict.items():
@@ -111,7 +119,7 @@ def generate_parfiles(parfile_dict=None, basename_parfile=None, output_dir=None)
         parfile_path = (output_dir / parfile_name).resolve()
         parfiles.append(str(parfile_path))
         f90nml.write(parfile_dict, parfile_path, force=True)
-    print('Parfiles generated, saved to {}'.format(output_dir))
+    pylboLogger.info('parfiles generated, saved to {}'.format(output_dir))
     return parfiles
 
 def _init_worker():
@@ -120,6 +128,26 @@ def _init_worker():
 def _activate_worker(parfile):
     cmd = ['./legolas', '-i', str(parfile)]
     return subprocess.call(cmd)
+
+def _terminate_workers():
+    # Note: simply giving an interrupt terminates ONLY the python processes,
+    # but still keeps the legolas calls running since those are done using subprocess.
+    # The following code first terminates all child processes (legolas), then the parents (workers)
+    pylboLogger.error('interrupting processes...')
+    for process in multiprocessing.active_children():
+        pid = process.pid
+        pylboLogger.error('terminating PID: {} -- {}'.format(pid, process.name))
+        parent = psutil.Process(pid)
+        children = parent.children(recursive=True)
+        for child in children:
+            pylboLogger.error('terminating child process {} -- {}'.format(child.pid, child.name()))
+            child.kill()
+        gone, alive = psutil.wait_procs(children, timeout=2)
+        for killed_proc in gone:
+            pylboLogger.error('{}'.format(str(killed_proc)))
+        parent.kill()
+        parent.wait(timeout=2)
+    pylboLogger.critical('all Legolas processes terminated.')
 
 def run_legolas(parfiles=None, remove_parfiles=False, nb_cpus=1):
     def update_pbar(*args):
@@ -132,50 +160,46 @@ def run_legolas(parfiles=None, remove_parfiles=False, nb_cpus=1):
     owd = os.getcwd()
     # change to source directory
     os.chdir(LEGOLAS_DIR)
-    # initialise progressbar and multiprocessing pool
-    pbar = tqdm(total=len(parfiles), unit='')
-    pbar.set_description('Running Legolas [CPUs={}]'.format(nb_cpus))
-    pool = multiprocessing.Pool(processes=nb_cpus, initializer=_init_worker)
-    try:
-        # start multiprocessing pool
-        for parfile in parfiles:
-            pool.apply_async(_activate_worker, args=(parfile,), callback=update_pbar)
-        pool.close()
-        pool.join()
-        pbar.close()
-    except KeyboardInterrupt:
-        pbar.set_description('INTERRUPTED')
-        pbar.update(len(parfiles))
-        pbar.close()
-        # Note: simply calling pool.terminate() terminates ONLY the python processes,
-        # but still keeps the legolas calls running since those are done using subprocess.
-        # The following code first terminates all child processes (legolas), then the parents (workers)
-        print('\nCaught KeyboardInterrupt:')
-        for process in multiprocessing.active_children():
-            pid = process.pid
-            print('  Terminating ID: {} -- {}'.format(pid, process.name))
-            parent = psutil.Process(pid)
-            children = parent.children(recursive=True)
-            for child in children:
-                print('    Terminating child process {} -- {}'.format(child.pid, child.name()))
-                child.kill()
-            gone, alive = psutil.wait_procs(children, timeout=2)
-            for killed_proc in gone:
-                print('    {}'.format(str(killed_proc)))
-            parent.kill()
-            parent.wait(timeout=2)
-        pool.terminate()
-        print('All processes terminated.\n')
+    # don't use multiprocessing if there is only one job
+    if len(parfiles) == 1:
+        parfile = parfiles[0]
+        pylboLogger.info("running Legolas...")
+        try:
+            _activate_worker(parfile)
+        except KeyboardInterrupt:
+            _terminate_workers()
+            exit(1)
+    else:
+        # initialise progressbar and multiprocessing pool
+        pbar = tqdm(total=len(parfiles), unit='')
+        pbar.set_description('Running Legolas [CPUs={}]'.format(nb_cpus))
+        pool = multiprocessing.Pool(processes=nb_cpus, initializer=_init_worker)
+        try:
+            # start multiprocessing pool
+            for parfile in parfiles:
+                pool.apply_async(_activate_worker, args=(parfile,), callback=update_pbar)
+            pool.close()
+            pool.join()
+            pbar.close()
+        except KeyboardInterrupt:
+            pbar.set_description('INTERRUPTED')
+            pbar.update(len(parfiles))
+            pbar.close()
+            _terminate_workers()
+            pool.terminate()
+            pool.join()
+            exit(1)
+    pylboLogger.info("all runs completed")
     # change back to original directory
     os.chdir(owd)
     # remove parfiles if asked
     if remove_parfiles:
         for file in parfiles:
             os.remove(file)
-        print('Parfiles removed.')
+        pylboLogger.info('Parfiles removed.')
         # if parfile directory is empty, also remove that one
         try:
             Path.rmdir(LEGOLAS_PAR)
-            print('{} was empty, so also removed the directory.'.format(LEGOLAS_PAR))
+            pylboLogger.info('{} was empty, so also removed the directory.'.format(LEGOLAS_PAR))
         except OSError:
             pass
